requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.1
tiktoken>=0.9.0 